+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 1.1.0
-- Added the [Priority Queue](docs/priority_queue.md) data structure
+- Added the [Priority Queue](docs/spsc/priority_queue.md) data structure
 
 ## 1.0.2
 - Made all class members accidentally public private